refactor: move the settings-change reaction into ApplySettingsChangeCommand
